Encoder aus for Uni-C

+++ dcg-panel.c
@@ -31,8 +31,8 @@
 #define NDEBUG
 #include "debug.h"
 
-int8_t EncDiff;
-int8_t EncDiffSpeed;
+int16_t EncDiff;
+int16_t EncDiffSpeed;
 
 uint8_t Mode;       // Mode = 0: Modification of the set value itself (with ModifyTimer running)
                     // Mode = 0: Display of measured value (in voltage / current mode (without ModifyTimer running)
